Divide the project in two parts
Daemon and client.
Daemon part will load pages and return HTML. This part will keep running
in the background to avoid the CPU load and time needed to load and
startup this part before each page load.
Client part will consist of a simple class that will catch argv, send it
to the daemon via socket-ipc and wait for the daemon to return (via
socket-ipc) the HTML. Then output it via stdout and quit.
This part should be small and quick (as it will be launched on each page
request). Right now it will
be based on QtCore, but maybe a plain C++ port will be made (if startup
time isn't fast enough).